Teacher_User_Managemet
n this code, we define two classes: User and Teacher. The User class represents a user with a username and password, while the Teacher class represents a teacher user with additional attributes for their first name, last name, and subject they teach. The Teacher class inherits from the User class.

We also define a User Manager class, which has a list of users and methods for adding and removing users, logging in a user with a given username and password, and getting a list of teachers who teach a given subject.